refactor(web): move clerk-config to shared directory
